+++ compiler/native/src/classes.cpp
@@ -1,0 +1,5 @@
+#include "./class/array.cpp"
+#include "./class/function.cpp"
+#include "./class/native.cpp"
+#include "./class/object.cpp"
+#include "./class/string.cpp"
