+++ compiler/native/src/classes.h
@@ -1,0 +1,6 @@
+#pragma once
+#include "./class/array.h"
+#include "./class/function.h"
+#include "./class/native.h"
+#include "./class/object.h"
+#include "./class/string.h"
