+++ compiler/native/src/values.h
@@ -1,0 +1,10 @@
+#pragma once
+#include "enum.h"
+
+namespace NJS::Value
+{
+	NJS::VAR undefined;
+	NJS::VAR null;
+	NJS::VAR Infinity;
+	NJS::VAR NaN;
+} // namespace NJS::Value
