+++ compiler/native/src2/types/base.h
@@ -1,0 +1,7 @@
+#include <vector>
+
+using namespace NJS;
+
+class NJS::CLASS::BaseObject {
+
+}
